+++ HISTORY.md
@@ -7,6 +7,8 @@
 - feat: added option to disable mount ownership
 - feat: added option to force mount ownership only in areas
   - player must be in an area where he has `area_addplayer` permission
+- fix: remove on hit debug message broadcast
+- fix: theft messages not properly removed for each player
 
 ## [0.4.0] - 2026-02-05 | Event logging (moved from DiscordConnect)